separate out the display commands from chessboard
