+++ ConsoleApp9/ChessBoard.cs
@@ -4,15 +4,17 @@ namespace Chessapp;
 
 public class ChessBoard
 {
-    private readonly Dictionary<string, IPiece?> _pieces = new();
-    private readonly List<(IPiece, (int, int))> _moves = new();
+    public readonly Dictionary<string, IPiece?> _pieces = new();
+    public readonly List<(IPiece, (int, int))> _moves = new();
     private readonly IPiece _blueKing, _greenKing;
+    private DisplayChessBoard _displayChessBoard;
 
     /// <summary>
     /// Constructs a GameState in a traditional chess layout.
     /// </summary>
     public ChessBoard()
     {
+        _displayChessBoard = new DisplayChessBoard(this);
         for (int i = 1; i <= 8; i++)
         {
             var pawnPiece = new PawnPiece("p" + i, PieceColor.Blue, (1, i - 1));
@@ -20,8 +22,10 @@ public class ChessBoard
             var piece = new PawnPiece("P" + i, PieceColor.Green, (6, i - 1));
             _pieces["P" + i] = piece;
         }
+
         _blueKing = _pieces["k1"] = new KingPiece("k1", PieceColor.Blue, (0, 3));
         _greenKing = _pieces["K1"] = new KingPiece("K1", PieceColor.Green, (7, 4));
+
         _pieces["q1"] = new QueenPiece("q1", PieceColor.Blue, (0, 4));
         _pieces["Q1"] = new QueenPiece("Q1", PieceColor.Green, (7, 3));
 
@@ -39,6 +43,7 @@ public class ChessBoard
         _pieces["R1"] = new RookPiece("R1", PieceColor.Green, (7, 0));
         _pieces["r2"] = new RookPiece("r2", PieceColor.Blue, (0, 7));
         _pieces["R2"] = new RookPiece("R2", PieceColor.Green, (7, 7));
+        
     }
 
     private void SetPiece((int row, int col) pos, IPiece? piece)
@@ -141,19 +146,7 @@ public class ChessBoard
     /// </summary>
     public void PrintBoard()
     {
-        Console.Clear();
-        DisplayLegend();
-        Console.WriteLine();
-
-        for (int row = 0; row < 8; row++)
-        {
-            DisplayRow(row);
-        }
-        DisplayColumns();
-        Console.WriteLine();
-
-        DisplayCapturedPieces();
-        DisplayPlayerTurn();
+       _displayChessBoard.PrintBoard();
     }
 
     /// <summary>
@@ -162,98 +155,6 @@ public class ChessBoard
     /// </summary>
     internal void DisplayPossibleMoves(List<(int, int)>? moves)
     {
-        (int left, int top) = Console.GetCursorPosition();
-        Console.ForegroundColor = ConsoleColor.Red;
-        if (moves != null)
-            foreach ((int row, int col) in moves)
-            {
-                Console.SetCursorPosition(col * 8, (row * 2) + 3);
-                Console.Write("|  XX  |");
-            }
-
-        Console.ResetColor();
-        Console.SetCursorPosition(left, top);
-    }
-
-    /// <summary>
-    /// Displays the bottom row of columns under the board.
-    /// </summary>
-    private static void DisplayColumns()
-    {
-        string[] columns = { "A", "B", "C", "D", "E", "F", "G", "H" };
-        Console.ForegroundColor = ConsoleColor.DarkYellow;
-        foreach (string letter in columns)
-        {
-            Console.Write("   {0}    ", letter);   
-        }
-        Console.ResetColor();
-    }
-
-    /// <summary>
-    /// Displays the legend at the top of the screen.
-    /// </summary>
-    private static void DisplayLegend()
-    {
-        Console.WriteLine("Legend: Use capital letters if you're green, lowercase if you're blue");
-        Console.WriteLine("R = Rook, N = Knight, P = Pawn, K = King, B = Bishop, Q = Queen");
-    }
-
-    /// <summary>
-    /// Displays a list of captured pieces.
-    /// </summary>
-    private void DisplayCapturedPieces()
-    {
-        Console.WriteLine("-------------------------------------------------------------------------");
-        Console.ForegroundColor = ConsoleColor.Red;
-        Console.WriteLine($"List of Pieces Captured: {string.Join(", ", _pieces.Values.Where(p => p is { IsPieceCaptured: true }).Select(p => p?.Symbol))}");
-        Console.ResetColor();
-        Console.WriteLine("-------------------------------------------------------------------------");
-    }
-
-    /// <summary>
-    /// Displays which player's turn it is.
-    /// </summary>
-    private void DisplayPlayerTurn()
-    {
-        PieceColor player = GetActivePlayer();
-        Console.ForegroundColor = player == PieceColor.Blue ? ConsoleColor.Cyan : ConsoleColor.Green;
-        Console.WriteLine($"It's {player}'s turn");
-        Console.ResetColor();
-    }
-
-    /// <summary>
-    /// Given a row and column, displays that cell on the screen.
-    /// </summary>
-    private void DisplayCell(int row, int col)
-    {
-        string symbol = "  ";
-        var isPiecePresent = _pieces.Values
-            .FirstOrDefault(x => x != null && x.Position.col == col && x.Position.row == row);
-        if (isPiecePresent != null)
-        {
-            symbol = isPiecePresent.Symbol;
-            Console.ForegroundColor = isPiecePresent.Color == PieceColor.Blue ? ConsoleColor.Cyan : ConsoleColor.Green;
-        }
-        Console.Write($"|  {symbol}  |");
-        Console.ResetColor();
-    }
-
-    /// <summary>
-    /// Given a row, displays that row on the screen
-    /// </summary>
-    private void DisplayRow(int row)
-    {
-        for (int col = 0; col < 8; col++)
-        {
-            DisplayCell(row, col);
-        }
-        Console.WriteLine();
-        Console.ForegroundColor = ConsoleColor.Yellow;
-        Console.Write(row + 1);
-        Console.ForegroundColor = ConsoleColor.DarkMagenta;
-        Console.Write("-----------------------------------------------------------------");
-        Console.ForegroundColor = ConsoleColor.Yellow;
-        Console.WriteLine(row + 1);
-        Console.ResetColor();
+        _displayChessBoard.DisplayPossibleMoves(moves);
     }
 }
